feat: /recommendページのシェア画像を統計入りの専用カードに刷新（LINE画像の直リンクをやめる）
これまでタグページ(/recommend/{tag})の og:image はランキング1位の部屋アイコン
（LINE CDN の画像URL直リンク）で、SNSでシェアしたときに「何のページか」が
伝わらず見栄えも良くなかった。/oc/{id}/card と同様の動的OGP画像
（/recommend/{tag}/card・1200x630 PNG・summary_large_image）に差し替える。

カードの内容: タグ見出し＋「人気・活発な部屋ランキング」、掲載部屋数と
メンバー合計、ランキング上位5部屋の円形アイコン（順位バッジ・メンバー数付き）、
サイト名・更新頻度・ドメイン。ja/tw/th の3ロケール対応（翻訳キー追加）。

実装:
- OcCardImageGenerator の多言語テキスト・絵文字・円形アイコン描画機構を
  AbstractCardImageGenerator に抽出し、oc用・recommend用の両カードで共用
- PNG送出/デフォルト画像フォールバックも OgCardHttpResponder として共通化
- タイトル省略(…)のバグ修正: 1行が1テキストランのとき行全体が消えて
  「…」だけになっていたのを、文字単位で末尾を削る方式に変更
- 生成は /oc カードと同じ ConcurrentRequestGuard('og-card') で 1IP×同時1本に
  制限し、エッジ(CF)キャッシュ12h＋日付クエリで日次ローテ

Committed from: vm:/home/user/Open-Chat-Graph
