+++ app/Controllers/Pages/RecommendOpenChatPageController.php
@@ -160,7 +160,11 @@ class RecommendOpenChatPageController
         $count = min($recommend->getCount(), AppConfig::LIST_LIMIT_RECOMMEND);
         $headline = sprintfT('「%s」のオープンチャット｜人気・活発な部屋ランキング', $tag);
         $_meta->setTitle($headline);
-        $_meta->setImageUrl(imgUrl($recommendList[0]['img_url']));
+        // og:image は統計焼き込みの動的カード（/recommend/{tag}/card）。以前は1位の部屋アイコン
+        // （LINE CDN直リンク）で、シェア時の見栄え・情報量に欠けた。日付クエリでSNS側の
+        // クロールキャッシュを1日単位で更新させる。thumbnail(検索用)は従来どおり1位の部屋アイコン
+        $_meta->setImageUrl(url('recommend/' . urlencode($tag) . '/card') . '?d=' . date('Ymd'))
+            ->setTwitterCard('summary_large_image');
         $_meta->thumbnail = imgPreviewUrl($recommendList[0]['img_url']);
 
         // 表示中の上位部屋を ItemList として同梱（AI検索・検索エンジンの機械可読引用向け）
