+++ storage/translation.json
@@ -348,6 +348,22 @@
         "tw": "「%s」的LINE社群｜熱門・活躍社群排行榜",
         "th": "OpenChat \"%s\" - อันดับห้องยอดนิยมและคึกคัก"
     },
+    "「%s」のオープンチャット": {
+        "tw": "「%s」的LINE社群",
+        "th": "OpenChat \"%s\""
+    },
+    "人気・活発な部屋ランキング": {
+        "tw": "熱門・活躍社群排行榜",
+        "th": "อันดับห้องยอดนิยมและคึกคัก"
+    },
+    "%1$s部屋を掲載・メンバー合計 %2$s人": {
+        "tw": "收錄 %1$s 個社群・成員合計 %2$s 人",
+        "th": "แสดง %1$s ห้อง・รวมสมาชิก %2$s คน"
+    },
+    "人": {
+        "tw": "人",
+        "th": " คน"
+    },
     "いま伸びている「%s」のオープンチャット": {
         "tw": "正在成長的「%s」LINE社群",
         "th": "OpenChat \"%s\" ที่กำลังเติบโต"
